+++ tempest_fastapi_sdk/settings/__init__.py
@@ -16,6 +16,7 @@ from tempest_fastapi_sdk.settings.mixins import AuthSettings as AuthSettings
 from tempest_fastapi_sdk.settings.mixins import CORSSettings as CORSSettings
 from tempest_fastapi_sdk.settings.mixins import DatabaseSettings as DatabaseSettings
 from tempest_fastapi_sdk.settings.mixins import EmailSettings as EmailSettings
+from tempest_fastapi_sdk.settings.mixins import GenAISettings as GenAISettings
 from tempest_fastapi_sdk.settings.mixins import JWTSettings as JWTSettings
 from tempest_fastapi_sdk.settings.mixins import LogSettings as LogSettings
 from tempest_fastapi_sdk.settings.mixins import MinIOSettings as MinIOSettings
@@ -36,6 +37,7 @@ __all__: list[str] = [
     "CORSSettings",
     "DatabaseSettings",
     "EmailSettings",
+    "GenAISettings",
     "JWTSettings",
     "LogSettings",
     "MinIOSettings",